Restructure cast trait modules
